# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "jump.rs",
    "description": "Wake-on-LAN API for managing and waking network devices",
    "license": {
      "name": "MIT"
    },
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login with username and password",
        "description": "Authenticates a user and returns a session cookie.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get current authenticated user",
        "description": "Returns information about the currently authenticated user.",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Current user info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          },
          {
            "basic_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response returned by all API endpoints",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Invalid MAC address format: xyz"
          },
          "status": {
            "type": "string",
            "description": "Always \"error\" for error responses",
            "example": "error"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "example": "password123"
          },
          "username": {
            "type": "string",
            "example": "admin"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": [
          "status",
          "username"
        ],
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "username": {
            "type": "string",
            "example": "admin"
          }
        }
      },
      "MeResponse": {
        "type": "object",
        "required": [
          "username"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          }
        }
      },
...
    }
  },
...
}